% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-setters.R
\name{abun<-}
\alias{abun<-}
\alias{abun<-,omic-method}
\alias{abun<-,omics-method}
\title{Set Abundance Data}
\usage{
abun(object) <- value
}
\arguments{
\item{object}{An \code{omic} or \code{omics} object.}

\item{value}{The new abundance data to be set.}
}
\value{
The \code{omic} or \code{omics} object with updated abundance data.
}
\description{
This setter function allows you to update the abundance data for \code{omic} objects
and each \code{omic} object within a \code{omics}. The abundance data must be a numeric matrix
for \code{omic} objects. For \code{omics} objects, the abundance data should be a named list
of numeric matrices corresponding to each \code{omic} object within the list.
}
