% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-check-assign.R
\name{are_list_assign}
\alias{are_list_assign}
\title{Check list in assign methods for omics}
\usage{
are_list_assign(object, value)
}
\arguments{
\item{object}{An \code{omics} object.}

\item{value}{A named list intended to be assigned to the \code{omics}.}
}
\description{
Internal function to validate that a list provided to an
assign/setter method is coherent with an \code{omics} object.
}
\keyword{internal}
