% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_omics.R
\name{as.list.omics}
\alias{as.list.omics}
\title{Convert an omics Object to a List}
\usage{
\method{as.list}{omics}(x, ...)
}
\arguments{
\item{x}{An \code{omics} object to be converted to a list.}

\item{...}{Not used.}
}
\value{
A list of \code{omic} objects.
}
\description{
Convert an omics Object to a List
}
