% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-getters.R
\name{comm}
\alias{comm}
\alias{comm,omic-method}
\alias{comm,omics-method}
\title{Get Community Detection Results}
\usage{
comm(object)
}
\arguments{
\item{object}{An \code{omic} or \code{omics} object.}
}
\value{
The community detection result object stored in the \code{comm} slot for an \code{omic} object,
or a list of such objects for an \code{omics} object, each representing the community
detection results of a contained \code{omic} object.
}
\description{
Retrieves the community detection results stored in the \code{comm} slot of an \code{omic} object
or each \code{omic} object within an \code{omics}. These results are typically derived from network
analysis methods and indicate the grouping or clustering of taxa into communities based on their
network interactions.
}
