% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-links.R
\name{deselect_link}
\alias{deselect_link}
\alias{deselect_link,omic-method}
\alias{deselect_link,omics-method}
\title{Deselect Links in an omic Object}
\usage{
deselect_link(object)
}
\arguments{
\item{object}{An \code{omic} or \code{omics} object.}
}
\value{
The same object, with no \code{selected_links} attribute.
}
\description{
Clears the \code{selected_links} attribute from an \code{omic} or \code{omics} object,
effectively removing any current link selection.
}
