% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_omic_validity.R
\docType{data}
\name{.OMICS_RESERVED_COLNAMES}
\alias{.OMICS_RESERVED_COLNAMES}
\title{Assert: no reserved keywords in column names}
\format{
An object of class \code{character} of length 9.
}
\usage{
.OMICS_RESERVED_COLNAMES
}
\description{
Assert: no reserved keywords in column names
}
\keyword{internal}
