% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-links.R
\name{get_grouped_link}
\alias{get_grouped_link}
\alias{get_grouped_link,omic-method}
\alias{get_grouped_link,omics-method}
\title{Retrieve Link Grouping in an omic Object}
\usage{
get_grouped_link(object)

\S4method{get_grouped_link}{omic}(object)

\S4method{get_grouped_link}{omics}(object)
}
\arguments{
\item{object}{An \code{omic} or \code{omics} object.}
}
\value{
For an \code{omic}, a numeric vector indicating the group ID of each link.
For an \code{omics}, a list of numeric vectors, one per sub-object.
}
\description{
Retrieves the integer group IDs (or list of group IDs for each sub-object)
previously assigned by \code{\link[=group_link]{group_link()}}.
}
\seealso{
\code{\link[=group_link]{group_link()}} for assigning these group IDs, and \code{\link[=ungroup_link]{ungroup_link()}}
for removing them.
}
