% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-links.R
\name{get_selected_links}
\alias{get_selected_links}
\alias{get_selected_links,omic-method}
\alias{get_selected_links,omics-method}
\title{Get Selected Links from an \code{omic} Object}
\usage{
get_selected_links(object)
}
\arguments{
\item{object}{An \code{omic} or \code{omics} object.}
}
\value{
A numeric vector (or named list of numeric vectors, for \code{omics})
containing the indices of selected links.
}
\description{
Retrieves the numeric indices of links previously selected by \code{\link[=select_link]{select_link()}}.
}
