% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-check-assign.R
\name{is_assign_omics_link_tbl}
\alias{is_assign_omics_link_tbl}
\title{Check Tibble in Assign Methods for omics links}
\usage{
is_assign_omics_link_tbl(object, value)
}
\arguments{
\item{object}{An \code{omics} object.}

\item{value}{A tibble/data frame with columns \code{omic},\code{from},\code{to} and \code{link_id}
and the additional information}
}
\value{
\code{invisible(TRUE)} if all checks pass; otherwise throws an error.
}
\description{
Internal validator for a tibble/data frame replacing all links
metadata across an \code{omics} object. Requires \code{omic},\code{from},\code{to} and \code{link_id}
columns and forbids row names.
}
\keyword{internal}
