% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-merge.R
\name{merge_omics}
\alias{merge_omics}
\title{Merge Multiple \code{omic} or \code{omics} Objects}
\usage{
merge_omics(
  ...,
  abun = NULL,
  rela = NULL,
  norm = NULL,
  meta = NULL,
  taxa = NULL,
  netw = NULL,
  comm = NULL
)
}
\arguments{
\item{...}{\code{omic} objects, \code{omics} objects, or lists of such objects to be merged.}

\item{abun}{A function to merge the 'abun' (abundance) slots from the provided \code{omic} objects, or \code{NULL} if no merging is required for this slot.}

\item{rela}{A function to merge the 'rela' (relative abundance) slots from the provided \code{omic} objects, or \code{NULL} if no merging is required for this slot.}

\item{norm}{A function to merge the 'norm' (normalized abundance) slots from the provided \code{omic} objects, or \code{NULL} if no merging is required for this slot.}

\item{meta}{A function to merge the 'meta' (metadata) slots from the provided \code{omic} objects, or \code{NULL} if no merging is required for this slot.}

\item{taxa}{A function to merge the 'taxa' (taxonomic data) slots from the provided \code{omic} objects, or \code{NULL} if no merging is required for this slot.}

\item{netw}{A function to merge the 'netw' (network) slots from the provided \code{omic} objects, or \code{NULL} if no merging is required for this slot.}

\item{comm}{A function to merge the 'comm' (community structure) slots from the provided \code{omic} objects, or \code{NULL} if no merging is required for this slot.}
}
\value{
A new \code{omic} or \code{omics} object containing the merged data from the input objects. Each slot of the returned object is the result of the merging function applied to that slot, or remains \code{NULL} if no function was specified for that slot.
}
\description{
Merges multiple \code{omic} or \code{omics} objects into a single \code{omic} or \code{omics} object by merging their respective slots according to user-specified functions. This method provides flexibility to define how each slot is combined, which is essential because a default merging behavior for such complex data structures is not feasible.
}
\details{
The function allows specifying custom merging strategies for each slot:
\itemize{
\item \strong{Abundance data (\code{abun})}: Functions like \code{dplyr::bind_cols} or \code{dplyr::bind_rows} might be appropriate.
\item \strong{Metadata (\code{meta})}: Merging can be handled by functions such as \code{dplyr::full_join} or \code{dplyr::bind_rows}.
\item \strong{Network structures (\code{netw})}: These might be combined using \code{igraph::union} or similar graph-specific functions.
This approach ensures flexibility in handling the complex and varied data stored within \code{omic} objects.
}

The method checks for consistency among \code{omics} objects. If merging \code{omics} objects, it first ensures that all have identical names in the 'omics' slot. Mismatched names will cause the function to stop and issue an error.
}
