% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-setters.R
\name{meta<-}
\alias{meta<-}
\alias{meta<-,omic-method}
\alias{meta<-,omics-method}
\title{Update Sample Metadata for omic and omics Objects}
\usage{
meta(object) <- value
}
\arguments{
\item{object}{An \code{omic} or \code{omics} object to be updated.}

\item{value}{The new sample metadata to set, either a dataframe or a named list
of dataframes as per the object type.}
}
\value{
The updated \code{omic} or \code{omics} object with new sample metadata.
}
\description{
This function sets the sample metadata for \code{omic} objects and also updates
each \code{omic} object within a \code{omics}. The metadata for \code{omic} objects must be
provided as a dataframe. For \code{omics} objects, the metadata should be either
a named list of dataframes corresponding to each \code{omic} object within the list,
or a single dataframe containing a 'omic' and 'sample_id' columns to split the
metadata accordingly.
}
