% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mutate.R
\name{mutate_meta}
\alias{mutate_meta}
\alias{mutate_meta,omic-method}
\alias{mutate_meta,omics-method}
\title{Modify and Augment \code{omic} Objects by Transforming the \code{meta} Slot}
\usage{
mutate_meta(object, ..., .ungroup = FALSE)
}
\arguments{
\item{object}{An \code{omic} or \code{omics} object.
The function targets the \code{sample} slot, which contains metadata for each sample.}

\item{...}{Dynamic expressions or functions to be applied to the data.
These expressions can manipulate both abundance data (e.g., 'abun', 'rela', 'norm') and
metadata within the sample slot. This allows for a comprehensive data transformation
experience that supports all standard and custom \code{tidyverse} manipulation techniques.}

\item{.ungroup}{Logical, default \code{FALSE}. If \code{TRUE}, remove the omic grouping
from the object at the end (i.e., drop the grouping attribute).}
}
\value{
Returns the \code{omic} or \code{omics} object with updated \code{sample} slots reflecting the applied transformations.
All other structures within the object remain unchanged, ensuring that only the targeted metadata is modified.
}
\description{
This function dynamically manipulates the \code{sample} slot within \code{omic} or \code{omics} objects,
applying user-defined transformations. It leverages the full suite of \code{tidyverse} tools, particularly
\code{dplyr}, to enable powerful and flexible data transformations.
}
\details{
The function is designed to integrate seamlessly with the \code{tidyverse}, allowing users
to utilize familiar and potent data manipulation verbs such as \code{mutate}, \code{filter}.
It supports using any \code{tidyverse}-compatible expressions, including conditional operations,
summarizations, and complex transformations involving both abundance and metadata fields.
This flexibility makes it particularly useful for ecological and biological data analysis,
where combining different data types and conditions is common.

\if{html}{\out{<div class="sourceCode">}}\preformatted{     ### Keywords in `omic` and `omics`:
     - **abun, rela, norm**: Slots within `omic` objects that store abundance data, which can be
       directly manipulated or used in conjunction with metadata to perform advanced analyses.
     - **sample_id**: An essential identifier used to uniquely reference individual samples within an `omic` object. 
     - **omic**: Used exclusively within `omics` objects to differentiate between multiple `omic` objects 
       contained in the list.
}\if{html}{\out{</div>}}
}
