% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-base-methods.R
\name{ncomm}
\alias{ncomm}
\alias{ncomm,omic-method}
\alias{ncomm,omics-method}
\title{Get Number of Communities}
\usage{
ncomm(object)
}
\arguments{
\item{object}{An object of class \code{omic} or \code{omics}.}
}
\value{
Integer (for \code{omic}) or named integer vector (for \code{omics}).
}
\description{
Returns the number of detected communities in an \code{omic} or each element of an \code{omics}.
Returns \code{0L} when communities are absent.
}
