% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-setters.R
\name{netw<-}
\alias{netw<-}
\alias{netw<-,omic-method}
\alias{netw<-,omics-method}
\title{Set Network Data}
\usage{
netw(object) <- value
}
\arguments{
\item{object}{An \code{omic} or \code{omics} object.}

\item{value}{The new network data to be set.}
}
\value{
The \code{omic} or \code{omics} object with updated network data.
}
\description{
This setter function updates the network data for \code{omic} objects and
each \code{omic} object within a \code{omics}. The network data must be an \code{igraph} object
for \code{omic} objects. For \code{omics} objects, the network data should be a named list of \code{igraph} objects
corresponding to each \code{omic} object within the list.
}
