% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-setters.R
\name{norm<-}
\alias{norm<-}
\alias{norm<-,omic-method}
\alias{norm<-,omics-method}
\title{Set Log-Transformed Abundance Data}
\usage{
norm(object) <- value
}
\arguments{
\item{object}{An \code{omic} or \code{omics} object.}

\item{value}{The new log-transformed abundance data to be set.}
}
\value{
The \code{omic} or \code{omics} object with updated log-transformed abundance data.
}
\description{
This setter function allows you to update the log-transformed abundance data for \code{omic} objects
and each \code{omic} object within a \code{omics}. The log-transformed abundance data must be a numeric matrix
for \code{omic} objects. For \code{omics} objects, the data should be a named list of numeric matrices
corresponding to each \code{omic} object within the list.
}
