% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis-base-methods.R
\name{nsample}
\alias{nsample}
\alias{nsample,omic-method}
\alias{nsample,omics-method}
\title{Get Number of Samples}
\usage{
nsample(object)
}
\arguments{
\item{object}{An object of class \code{omic} or \code{omics}.}
}
\value{
Integer (for \code{omic}) or named integer vector (for \code{omics}).
}
\description{
Returns the number of samples in an \code{omic} or \code{omics} object.
}
\details{
For an empty object, returns \code{0L}. For \code{omics}, returns a named integer
vector with one value per element.
}
