% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-base-methods.R
\name{ntaxa}
\alias{ntaxa}
\alias{ntaxa,omic-method}
\alias{ntaxa,omics-method}
\title{Get Number of Taxa}
\usage{
ntaxa(object)
}
\arguments{
\item{object}{An object of class \code{omic} or \code{omics}.}
}
\value{
Integer (for \code{omic}) or named integer vector (for \code{omics}).
}
\description{
Returns the number of taxa in an \code{omic} or \code{omics} object.
}
\details{
For an empty object, returns \code{0L}. For \code{omics}, returns a named integer
vector with one value per element.
}
