% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-merge.R
\name{omic_collapse}
\alias{omic_collapse}
\alias{omic_collapse,omics-method}
\title{Collapse Specified Combinations within an omics}
\usage{
omic_collapse(
  object,
  by = NULL,
  abun = NULL,
  rela = NULL,
  norm = NULL,
  meta = NULL,
  taxa = NULL,
  netw = NULL,
  comm = NULL
)
}
\arguments{
\item{object}{An \code{omics} containing multiple \code{omic} objects.}

\item{by}{A list specifying the combinations of \code{omic} object names to be merged. Each element in
the list should be a character vector containing the names of the \code{omic} objects to merge.
If NULL, all \code{omic} objects in the list are merged into a single \code{omic} object.}

\item{abun}{A function to merge the 'abun' (abundance) slots from the selected \code{omic} objects, or
NULL if no merging is required for this slot.}

\item{rela}{A function to merge the 'rela' (relative abundance) slots, or NULL if not required.}

\item{norm}{A function to merge the 'norm' (normalized abundance) slots, or NULL if not required.}

\item{meta}{A function to merge the 'meta' (metadata) slots, or NULL if not required.}

\item{taxa}{A function to merge the 'taxa' (taxonomic data) slots, or NULL if not required.}

\item{netw}{A function to merge the 'netw' (network) slots, or NULL if not required.}

\item{comm}{A function to merge the 'comm' (community structure) slots, or NULL if not required.}
}
\value{
An \code{omics} containing the merged \code{omic} objects. Each entry in the returned \code{omics}
corresponds to a merged group as defined in the \code{by} parameter. The names of the entries
reflect the groups specified in \code{by} or are derived by concatenating the names of the merged
\code{omic} objects with dashes if \code{by} is unnamed.
}
\description{
Merges specific combinations of \code{omic} objects within an \code{omics} according to user-defined rules,
allowing for targeted merging of subsets within the list. This function facilitates focused analyses
by enabling the aggregation of related data into meaningful groups.
}
\details{
The function checks for the validity of names specified in the \code{by} parameter against the names
of \code{omic} objects in the \code{omics}. It ensures that all specified groups contain valid \code{omic}
object names and that each group's name is unique. This function is useful for scenarios where
specific subsets of data within an \code{omics} need to be aggregated based on certain criteria
or experimental conditions.

If \code{by} is NULL, a default merge of all \code{omic} objects in the list is performed, resulting in a
single \code{omic} object.

Example usage:
\dontrun{
  # Assuming list1 contains several omic objects named 'A', 'B', 'C', etc.
  result <- omic_collapse(list1, by = list(Group1 = c("A", "B"), Group2 = c("C")),
                           meta = function(x) dplyr::bind_rows(x),
                           netw = igraph::graph.union)
}
}
