% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-longer.R
\name{omic_longer}
\alias{omic_longer}
\alias{omic_longer,omic-method}
\alias{omic_longer,omics-method}
\title{Summarize \code{omic} and \code{omics} Objects in Long Format}
\usage{
omic_longer(object)
}
\arguments{
\item{object}{An \code{omic} or \code{omics} object.
For \code{omic}, this method consolidates sample and taxa information along with
abundance data and associated metadata into a long-format tibble. For \code{omics},
it applies the transformation to each \code{omic} object within the list and combines the results
into a single tibble, incorporating an additional column named \code{omic} that identifies the source
\code{omic} object.}
}
\value{
A tibble in a long format that consolidates all the relevant data fields under sample and taxon identifiers.
For \code{omics} objects, the tibble includes an extra column \code{omic} that indicates the source \code{omic} object.
}
\description{
Converts \code{omic} or \code{omics} objects into a longer format including abundances
(absolute, relative, and normalized) and metadata into a single structured tibble.
This transformation facilitates the use of tidyverse tools for
subsequent data manipulation and analysis.
}
\details{
The \code{omic_longer} method is tailored to prepare \code{omic} data for easy integration with tidyverse
workflows by:
\itemize{
\item Transforming data into a long format based on \code{sample_id} and \code{taxa_id}, making it more accessible for
various tidyverse functions such as \code{mutate}, \code{filter}, and \code{summarize}.
\item Aggregating both abundance metrics (e.g., \code{abun}, \code{rela}, \code{norm}) and metadata associated (e.g meta, taxa, comm_id)
with samples and taxa, which allows comprehensive data analysis within a single data frame structure.
\item In the context of \code{omics} objects, besides the standard data transformation applied to each \code{omic},
an additional column using the reserved keyword \code{omic} is included to keep track of the original \code{omic}
object each row of data pertains to, facilitating analysis across multiple datasets.
}
}
