% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_omics.R
\name{omics-access}
\alias{omics-access}
\alias{$,omics-method}
\alias{$<-,omics-method}
\title{Access or Assign Elements in an omics Object}
\usage{
\S4method{$}{omics}(x, name)

\S4method{$}{omics}(x, name) <- value
}
\arguments{
\item{x}{An object of class \code{omics}.}

\item{name}{The name of the \code{omic} object to access or assign within the \code{omics} list.}

\item{value}{An object of class \code{omic} to assign to the given name (used in \verb{$<-} only).}
}
\value{
For \code{$}, returns the \code{omic} object with the specified name.
For \verb{$<-}, returns the updated \code{omics} object with the new or modified entry.
}
\description{
These methods allow access to or assignment of \code{omic} objects within an \code{omics} object using the \code{$} operator.
}
