% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_omics.R
\docType{class}
\name{omics-class}
\alias{omics-class}
\title{Class omics}
\description{
\code{omics} is an S4 class designed to hold and manage a collection of \code{omic} objects,
providing a structured way to handle multiple NGS datasets. It ensures that all
elements within the list adhere to the \code{omic} class structure, making it useful for
analyses involving multiple datasets. This class allows for consistent and
type-safe manipulation of multiple \code{omic} objects.
}
\section{Slots}{

\describe{
\item{\code{omics}}{A list where each element is an instance of the \code{omic} class.
If the list is not empty, each \code{omic} object within the list must be named to facilitate
easy access and management.}
}}

\note{
For clarity, note that the class is named \code{omics} and it contains a slot named \code{omics} as well.
}
\section{Validity Checks}{

The class includes validity checks to ensure that, if not empty, all elements within the
\code{omics} slot are instances of the \code{omic} class and are named. These checks help maintain
integrity and usability, preventing errors during downstream analyses.
}

\seealso{
\code{\link[=omic]{omic}} for details on the \code{omic} class and its functionalities.
}
