% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-omics.R
\name{omics-methods}
\alias{omics-methods}
\alias{as.list.omics}
\alias{coerce,omics,list-method}
\alias{length,omics-method}
\alias{names,omics-method}
\alias{names<-,omics,character-method}
\title{General methods for \code{omics} objects}
\usage{
\method{as.list}{omics}(x, ...)

\S4method{length}{omics}(x)

\S4method{names}{omics}(x)

\S4method{names}{omics,character}(x) <- value
}
\arguments{
\item{x}{An object of class \code{omics}.}

\item{...}{Not used.}

\item{value}{A character vector for the replacement form \verb{names<-}.}
}
\value{
\itemize{
\item \code{as.list()} / \code{as(., 'list')}: a list of \code{omic} objects. \cr
\item \code{length()}: an integer with the number of elements. \cr
\item \code{names()}: a character vector with element names. \cr
\item \verb{names<-()}: the updated \code{omics} object.
}
}
\description{
Basic utilities to interact with \code{omics} objects: coerce to list, get length,
get and set names.
}
\section{Methods}{

\describe{
\item{\code{as.list(x)}}{Convert an \code{omics} object to a plain list of \code{omic} objects.}
\item{\code{as(x, 'list')}}{Coerce the \code{omics} object to a list.}
\item{\code{length(x)}}{Return the number of contained \code{omic} objects.}
\item{\code{names(x)}}{Return element names.}
\item{\code{names(x) <- value}}{Set element names.}
}
}

