% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_omics.R
\name{omics}
\alias{omics}
\title{Create an omics Object}
\usage{
omics(...)
}
\arguments{
\item{...}{Optional \code{omic} objects or a list of \code{omic} objects to be included in the \code{omics} object.}
}
\value{
An \code{omics} object containing any provided \code{omic} objects. If not empty,
each \code{omic} object must be named.
}
\description{
Constructs an \code{omics} object to encapsulate and manage a collection of \code{omic} objects.
This function ensures that all provided \code{omic} objects are validated and organized into a
structured list, facilitating easy access to multiple datasets. The constructor
supports the initialization of an empty list, individual \code{omic} objects, or a list of \code{omic} objects.
}
