% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-setters.R
\name{rela<-}
\alias{rela<-}
\alias{rela<-,omic-method}
\alias{rela<-,omics-method}
\title{Set rela Data}
\usage{
rela(object) <- value
}
\arguments{
\item{object}{An \code{omic} or \code{omics} object.}

\item{value}{The new rela data to be set.}
}
\value{
The \code{omic} or \code{omics} object with updated abundance data.
}
\description{
This setter function allows you to update the rela data for \code{omic} objects
and each \code{omic} object within a \code{omics}. The rela data must be a numeric matrix
for \code{omic} objects. For \code{omics} objects, the rela data should be a named list
of numeric matrices corresponding to each \code{omic} object within the list.
}
