% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-getters.R
\name{rela}
\alias{rela}
\alias{rela,omic-method}
\alias{rela,omics-method}
\title{Retrieve Relative Abundance Data Based on Taxonomic Information}
\usage{
rela(object, .fmt = "mat", .var = NULL, .fun = sum)
}
\arguments{
\item{object}{An \code{omic} or \code{omics} object containing relative abundance and informational data.}

\item{.fmt}{A character string specifying the output format of the result.
Possible choices are:
- "mat": returns a matrix
- "df": returns a data.frame
- "tbl": returns a tibble, where for \code{omic} objects, the row names of the relative abundance
matrix are moved into a new column named \code{sample_id}.
The default format is "mat".}

\item{.var}{A character string specifying the column name from the \code{taxa} slot to be used
for grouping and aggregation. If this parameter is omitted, the function will return
the original relative abundance data without any aggregation.}

\item{.fun}{A function to specify how the relative abundance data should be aggregated.
This can be any function that takes a numeric vector and returns a single number (e.g., sum, mean).
The default aggregation function is sum.}
}
\value{
Depending on the '.fmt' parameter:
- For an \code{omic} object, returns relative abundance data formatted as specified, aggregated
based on the provided column if specified.
- For an \code{omics} object, returns a list of such formatted data from each \code{omic}
object within the list, enabling batch processing and analysis of multiple datasets.
}
\description{
Retrieves and aggregates relative abundance data for an \code{omic} or \code{omics} object based on
a specified column in the \code{taxa} data frame. This function allows for custom aggregation
methods such as sum or mean. If no column is specified, it returns the original relative abundance data
in the selected format. Although \code{taxa} can contain various types of information, it is
commonly used to obtain relative abundances at different taxonomic ranks or other categorical variables.
}
