% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-split.R
\name{split_taxa}
\alias{split_taxa}
\alias{split_taxa,omic-method}
\title{Split an omic Object into an omics Based on Specified Columns}
\usage{
split_taxa(object, ...)
}
\arguments{
\item{object}{An \code{omic} object.}

\item{...}{Columns in the \code{taxa} data frame used to define subsets.
You can specify columns by name, using any of the selection methods supported by \code{dplyr::select()}.}
}
\value{
An \code{omics} object containing \code{omic} objects for each unique combination
of specified column values.
}
\description{
This method splits an \code{omic} object into subsets based on unique combinations
of values in specified columns from the \code{taxa} data frame. Each subset
corresponds to an \code{omic} object in the resulting \code{omics}, facilitating
analyses specific to each unique combination of metadata values.
}
\details{
The \code{split_taxa} function leverages \code{dplyr} functionality to identify unique combinations
of specified column values in the \code{taxa} data frame. It then creates a new \code{omic}
object for each combination, allowing for tailored analysis per group.

This function is particularly useful in exploratory data analysis and pre-processing stages
where data need to be examined or analyzed based on specific grouping variables.
}
\note{
This function depends on \code{dplyr} for data manipulation. Ensure that \code{dplyr} is installed
and loaded in your R session.
}
