% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-getters.R
\name{taxa}
\alias{taxa}
\alias{taxa,omic-method}
\alias{taxa,omics-method}
\title{Get Taxa Metadata Information}
\usage{
taxa(object, .fmt = "df", .collapse = FALSE)
}
\arguments{
\item{object}{An \code{omic} or \code{omics} object.}

\item{.fmt}{A character string specifying the output format of the result.
Possible choices are:
- "df": Returns the output as a \code{data.frame}.
- "tbl": Returns the output as a \code{tibble}. For \code{omic} objects, the the row names of
the abundance matrix are converted into a new column named \code{taxa_id}, ensuring alignment
with the reserved keyword in \code{omic-class}.}

\item{.collapse}{Logical, only for \code{omics}. If TRUE, ignore \code{.fmt} and return
a single tibble with all taxas row-bound together and an \code{omic} column
indicating the source object. Default: FALSE.}
}
\value{
The content of the \code{taxa} slot for \code{omic} or a list of such contents for \code{omics}.
}
\description{
Retrieves the taxa information stored in the \code{taxa} slot of an \code{omic} object
or for each \code{omic} object within an \code{omics}, with the option to format the output as
a \code{data.frame}, \code{tibble}, or a combined \code{tibble} for multiple omic objects.
}
