% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-links.R
\name{ungroup_link}
\alias{ungroup_link}
\alias{ungroup_link,omic-method}
\alias{ungroup_link,omics-method}
\title{Ungroup Links in an omic Object}
\usage{
ungroup_link(object)

\S4method{ungroup_link}{omic}(object)

\S4method{ungroup_link}{omics}(object)
}
\arguments{
\item{object}{An \code{omic} or \code{omics} object.}
}
\value{
The same object, with no \code{link_groups} attribute.
}
\description{
Removes any existing link grouping in an \code{omic} or \code{omics} object, clearing
the \code{link_groups} attribute so that link operations within \code{\link[=mutate_link]{mutate_link()}}
are no longer grouped.
}
