% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-omics.R
\name{omics-access}
\alias{omics-access}
\alias{$,omics-method}
\alias{$<-,omics,omic-method}
\alias{[[,omics-method}
\alias{[[<-,omics,ANY,ANY,omic-method}
\title{Access or assign \code{omic} elements in an \code{omics} object}
\usage{
\S4method{$}{omics}(x, name)

\S4method{$}{omics,omic}(x, name) <- value

\S4method{[[}{omics}(x, i, j, ...)

\S4method{[[}{omics,ANY,ANY,omic}(x, i, j) <- value
}
\arguments{
\item{x}{An \code{omics} object.}

\item{name}{Length-1 character: the element name (for \code{$}).}

\item{value}{An \code{omic} object to assign.}

\item{i}{Index or name (for \code{[[} / \verb{[[<-}).}

\item{j}{Ignored (signature compatibility).}

\item{...}{Not used.}
}
\value{
\itemize{
\item \code{$} / \code{[[}: the selected \code{omic} object. \cr
\item \verb{$<-} / \verb{[[<-}: the updated \code{omics} object.
}
}
\description{
List-like accessors using \code{$}, \verb{$<-}, \code{[[}, and \verb{[[<-}.
}
