% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_omics.R
\name{omics-access}
\alias{omics-access}
\alias{$,omics-method}
\alias{$<-,omics-method}
\title{Access or Assign Elements in an omics Object}
\usage{
\S4method{$}{omics}(x, name)

\S4method{$}{omics}(x, name) <- value
}
\description{
Access or Assign Elements in an omics Object
}
