% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-omics.R
\docType{class}
\name{omics-class}
\alias{omics-class}
\title{Class \code{omics} – a collection of \code{omic} objects}
\description{
\code{omics} is an S4 class to hold and manage a collection of \code{omic} objects,
enabling multi-dataset analyses with a consistent, type-safe interface.
}
\section{Slots}{

\describe{
\item{\code{omics}}{A list whose elements are instances of class \code{omic}.
If not empty, the list must be named and names must be unique.}
}}

\section{Validity checks}{

If not empty:
\itemize{
\item all elements must inherit from class \code{omic};
\item all elements must have non-empty names;
\item names must be unique.
}
}

\seealso{
\code{\link[=omic]{omic()}} for details about the single-omic class used as elements.
}
