% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-omics.R
\name{omics}
\alias{omics}
\title{Create an \code{omics} object}
\usage{
omics(...)
}
\arguments{
\item{...}{\code{omic} objects, or a \emph{list} of \code{omic} objects.}
}
\value{
An object of class \code{omics}. If not empty, elements must be named with
unique names.
}
\description{
Build an \code{omics} object to encapsulate a collection of \code{omic} objects.
Accepts multiple \code{omic} arguments or a single \emph{list} of \code{omic}.
}
